Update methodological_specification.md
Removed the error list as only 5 were listed instead of 15 and made reference that the explanations will be made in the technical spec and they are in the code. I've also added in that the method will continue processing even when errors occur and this was done for historical reasons, unlike the other methods in the SML.

These changes were agreed upon by DST and MQD in a meeting on the 16th Jan.